fix compilation with apache 2.4 on trusty fix warning

+++ src/RequestProcessor.cc
@@ -68,7 +68,7 @@ Commands::toDuplicate() {
         return false;
     }
 
-    return ((randNum % 100) < mDuplicationPercentage);
+    return ((randNum % 100) < static_cast<int>(mDuplicationPercentage));
 }
